requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
supabase
python-dotenv
pydantic-settings
sqlalchemy
alembic
psycopg[binary]
